Add navbar to recipe app
