--- app/helpers/recipes_food_helper.rb
@@ -1,2 +1,0 @@
-module RecipesFoodHelper
-end
