+++ app/helpers/recipes_foods_helper.rb
@@ -1,0 +1,2 @@
+module RecipesFoodsHelper
+end
