--- app/views/recipes_food/_recipes_food.html.erb
@@ -1,2 +1,0 @@
-<div id="<%= dom_id recipes_food %>">
-</div>
