--- app/views/recipes_food/new.html.erb
@@ -1,5 +1,0 @@
-<h1 class="text-center">New recipes food</h1>
-
-<%= render "form", recipes_food: @recipes_food %>
-
-<br>
