--- app/views/recipes_food/show.html.erb
@@ -1,10 +1,0 @@
-<p style="color: green"><%= notice %></p>
-
-<%= render @recipes_food %>
-
-<div>
-  <%= link_to "Edit this recipes food", edit_recipesFood_path(@recipes_food) %> |
-  <%= link_to "Back to recipes foods", recipesFoods_path %>
-
-  <%= button_to "Destroy this recipes food", @recipes_food, method: :delete %>
-</div>
