+++ app/views/recipes_foods/_recipes_food.html.erb
@@ -1,0 +1,2 @@
+<div id="<%= dom_id recipes_food %>">
+</div>
