+++ app/views/recipes_foods/edit.html.erb
@@ -1,0 +1,10 @@
+<h1>Editing recipes food</h1>
+
+<%= render "form", recipes_food: @recipes_food %>
+
+<br>
+
+<div>
+  <%= link_to "Show this recipes food", @recipes_food %> |
+  <%= link_to "Back to recipes foods", recipesFoods_path %>
+</div>
