+++ app/views/recipes_foods/new.html.erb
@@ -1,0 +1,5 @@
+<h1 class="text-center">New recipes food</h1>
+
+<%= render "form", recipes_food: @recipes_food %>
+
+<br>
